+++ Assets/Code/Scripts/UI/Modals/FundsDepositModal.cs
@@ -45,7 +45,7 @@ public class FundsDepositModal : Modal {
 
     [ClientCallback]
     private void OnDepositSubmitButtonClicked() {
-        uint depositValue = this.depositValueField.value;
+        int depositValue = (int)this.depositValueField.value;
         var localPlayer = BoardContext.Instance.GetLocalPlayer();
 
         if (depositValue <= 0) {
@@ -65,10 +65,10 @@ public class FundsDepositModal : Modal {
     }
 
     [Command(requiresAuthority = false)]
-    private void CmdDepositFunds(uint depositValue, BoardPlayer localPlayer) {
+    private void CmdDepositFunds(int depositValue, BoardPlayer localPlayer) {
         FundsHistoryEntry entry = new FundsHistoryEntry(depositValue, HistoryEntryType.DEPOSIT, "Einzahlung von " + localPlayer.PlayerName);
         BoardContext.Instance.fundsHistory.Add(entry);
-        BoardContext.Instance.UpdateFundsStat((int)depositValue);
+        BoardContext.Instance.UpdateFundsStat(depositValue);
         localPlayer.RemoveCoins(depositValue);
     }
 
@@ -84,9 +84,9 @@ public class FundsDepositModal : Modal {
         this.OnDepositAddButtonClicked(1000);
     }
 
-    private void OnDepositAddButtonClicked(uint amount) {
+    private void OnDepositAddButtonClicked(int amount) {
         if (this.depositValueField != null) {
-            this.depositValueField.value += amount;
+            this.depositValueField.value += (uint)amount;
         }
     }
